review(pr#142): address Copilot + accuracy fixes
Copilot:
- use imported `ReactNode` instead of the `React.ReactNode` global in the
  PathCard/Step helpers.
- quickstart heading + metadata now say "4 steps" (matches the 4 rendered
  steps).
- registry-scoped navbar keeps the "Download" link.
- add unit tests for the SDK's multicall decode helpers (vitest), mirroring
  the frontend coverage — the published package was untested.

Accuracy / robustness (Gemini, where correct):
- verifiedUntil returns uint64 (matches IdentityRegistry.sol + the frontend
  ABI) in the SDK ABI, the contract-reference page, and llms.txt.
- getRegistryInfo throws when a registry isn't registered in the factory
  (was returning a zero-value dummy); getRegistryInfos returns null for
  those entries, keeping the aligned-array contract.
- CLI `info` rejects an invalid registry address instead of silently
  falling back to the default.

+++ frontend/app/developers/contracts/page.tsx
@@ -9,7 +9,7 @@ export const metadata = {
 const INTERFACE = `interface IIdentityRegistry {
     // --- gating ---
     function isVerified(address wallet) external view returns (bool);
-    function verifiedUntil(address wallet) external view returns (uint256);
+    function verifiedUntil(address wallet) external view returns (uint64);
 
     // --- policy / metadata ---
     function owner() external view returns (address);
